[Refactor] 코딩 테스트 페이지 구조 개편 및 컴포넌트 분리
