--- src/pages/coding-test/components/CodeEditor.scss
@@ -1,45 +1,0 @@
-.code-editor-wrapper {
-  width: 100%;
-  height: 100%;
-  border-radius: 8px;
-  overflow: hidden;
-  box-sizing: border-box;
-
-  .code-editor-container {
-    width: 100%;
-    height: 100%;
-    overflow: auto;
-    box-sizing: border-box;
-
-    .cm-editor {
-      height: 100%;
-      width: 100%;
-
-      .cm-scroller {
-        overflow: auto;
-        font-family: 'Fira Code', monospace, 'Courier New';
-        font-size: 14px;
-        line-height: 1.5;
-        width: 100%;
-        scrollbar-width: none; /* Firefox */
-        &::-webkit-scrollbar {
-          display: none; /* Chrome, Safari, Opera */
-        }
-      }
-
-      .cm-content {
-        width: 100%;
-        box-sizing: border-box;
-        white-space: pre;
-        overflow-x: auto;
-      }
-
-      .cm-line {
-        width: 100%;
-        box-sizing: border-box;
-        white-space: pre;
-        overflow: visible;
-      }
-    }
-  }
-}
