+++ src/pages/coding-test/components/EditorPanel.scss
@@ -1,0 +1,146 @@
+@import '@/pages/coding-test/components/_variables.scss';
+
+.editor-panel {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: row;
+  background-color: $dark-bg-primary;
+
+  .main-editor-area {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    overflow: hidden;
+
+    .editor-header {
+      display: flex;
+      justify-content: flex-end;
+      align-items: center;
+      background-color: $dark-bg-secondary;
+      border-bottom: 1px solid $border-secondary;
+      height: $editor-header-height;
+      min-height: $editor-header-height;
+      flex-shrink: 0;
+      padding: 0 $spacing-sm 0 0;
+
+      .editor-tabs {
+        display: flex;
+        flex: 1;
+        overflow-x: auto;
+        padding-bottom: 0;
+        @include hide-scrollbar;
+      }
+
+      .editor-tab {
+        padding: $spacing-sm $spacing-lg;
+        background-color: $dark-bg-tertiary;
+        border-right: 1px solid $border-primary;
+        color: $text-tertiary;
+        white-space: nowrap;
+        display: flex;
+        align-items: center;
+        height: 100%;
+
+        &.active {
+          background-color: $dark-bg-active;
+          color: $text-light;
+        }
+      }
+
+      .language-selector-container {
+        display: flex;
+        align-items: center;
+        padding-right: $spacing-lg;
+      }
+    }
+
+    .language-selector {
+      background-color: $dark-bg-input;
+      color: $text-primary;
+      border: 1px solid $border-tertiary;
+      border-radius: $radius-sm;
+      padding: 0.3rem 0.6rem;
+      font-size: $font-base;
+      cursor: pointer;
+      outline: none;
+
+      &:hover {
+        background-color: $dark-bg-hover;
+      }
+
+      option {
+        background-color: $dark-bg-input;
+        color: $text-primary;
+      }
+    }
+
+    .editor-container {
+      width: 100%;
+      height: 100%;
+      overflow: hidden;
+      min-height: 0;
+
+      &.code-editor-wrapper {
+        border-radius: 8px;
+        box-sizing: border-box;
+
+        .code-editor-container {
+          width: 100%;
+          height: 100%;
+          overflow: auto;
+          box-sizing: border-box;
+
+          .cm-editor {
+            height: 100%;
+            width: 100%;
+
+            .cm-scroller {
+              overflow: auto;
+              font-family: 'Fira Code', monospace, 'Courier New';
+              font-size: 14px;
+              line-height: 1.5;
+              width: 100%;
+              scrollbar-width: none;
+              &::-webkit-scrollbar {
+                display: none;
+              }
+            }
+
+            .cm-content {
+              width: 100%;
+              box-sizing: border-box;
+              white-space: pre;
+              overflow-x: auto;
+            }
+
+            .cm-line {
+              width: 100%;
+              box-sizing: border-box;
+              white-space: pre;
+              overflow: visible;
+            }
+          }
+        }
+      }
+    }
+
+    .output-panel {
+      width: 100%;
+      height: 100%;
+      background-color: $dark-bg-secondary;
+      border-top: 1px solid $border-secondary;
+      padding: $spacing-md;
+      overflow-y: auto;
+
+      pre {
+        margin: 0;
+        font-family: 'Fira Code', monospace, 'Courier New';
+        font-size: $font-base;
+        color: $text-secondary;
+        white-space: pre-wrap;
+        word-break: break-all;
+      }
+    }
+  }
+}
