+++ src/pages/coding-test/components/Header.tsx
@@ -1,0 +1,35 @@
+import React from 'react';
+import './Header.scss';
+
+interface HeaderProps {
+  remainingTime: string;
+  isRunning: boolean;
+  handleRunCode: () => void;
+  handleSubmit: () => void;
+}
+
+const Header: React.FC<HeaderProps> = ({
+  remainingTime,
+  isRunning,
+  handleRunCode,
+  handleSubmit
+}) => {
+  return (
+    <div className="header">
+      <div className="header-timer">남은 시간 {remainingTime}</div>
+      <div className="header-title">9BACKPRO</div>
+      <div className="header-actions">
+        <button
+          className="run-button header-button"
+          onClick={handleRunCode}
+          disabled={isRunning}
+        >
+          {isRunning ? '실행중...' : '실행하기'}
+        </button>
+        <button className="submit-button header-button" onClick={handleSubmit}>제출하기</button>
+      </div>
+    </div>
+  );
+};
+
+export default Header;
