+++ src/pages/coding-test/components/ProblemPanel.scss
@@ -1,0 +1,87 @@
+@import '@/pages/coding-test/components/_variables.scss';
+
+.problem-panel {
+  width: 100%;
+  height: 100%;
+  background-color: $dark-bg-primary;
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+
+  .problem-header {
+    padding: $spacing-lg;
+    border-bottom: 1px solid $border-primary;
+    flex-shrink: 0;
+
+    .problem-tabs {
+      display: flex;
+      gap: $spacing-sm;
+      margin-bottom: $spacing-lg;
+      overflow-x: auto;
+      padding-bottom: 5px;
+
+      @include hide-scrollbar;
+
+      .problem-tab {
+        padding: $spacing-sm $spacing-lg;
+        background-color: $dark-bg-tertiary;
+        border: 1px solid $dark-bg-tertiary;
+        border-radius: $radius-xl;
+        font-size: $font-base;
+        cursor: pointer;
+        color: $text-primary;
+        white-space: nowrap;
+
+        &.active {
+          background-color: $dark-bg-active;
+          border-color: $dark-bg-active;
+        }
+      }
+    }
+
+    .problem-title {
+      font-size: $font-xl;
+      text-align: left;
+      margin: 0;
+      padding-left: $spacing-xs;
+      color: $text-primary;
+    }
+  }
+
+  .problem-details {
+    padding: $spacing-lg;
+    background-color: $dark-bg-primary;
+    flex: 1;
+    overflow-y: auto;
+    overflow-x: hidden;
+    min-height: 0;
+
+    h3 {
+      font-size: $font-md;
+      margin: $spacing-xl 0 $spacing-sm;
+      color: $text-secondary;
+
+      &:first-child {
+        margin-top: 0;
+      }
+    }
+
+    p {
+      margin: 0 0 $spacing-lg;
+      line-height: 1.5;
+      white-space: pre-wrap;
+      word-break: keep-all;
+    }
+
+    pre {
+      background-color: $dark-bg-secondary;
+      padding: $spacing-md;
+      border-radius: $radius-sm;
+      font-family: 'Fira Code', monospace, 'Courier New';
+      font-size: $font-base;
+      color: #f8f8f2;
+      white-space: pre;
+      overflow-x: auto;
+    }
+  }
+}
