+++ src/pages/coding-test/components/_variables.scss
@@ -1,0 +1,71 @@
+// 색상 변수
+$dark-bg-primary: #1e1e1e;
+$dark-bg-secondary: #252526;
+$dark-bg-tertiary: #2b2b2b;
+$dark-bg-active: #3a3d41;
+$dark-bg-input: #3c3c3c;
+$dark-bg-hover: #4a4a4a;
+
+$border-primary: #444;
+$border-secondary: #666;
+$border-tertiary: #555;
+
+$text-primary: #e0e0e0;
+$text-secondary: #cccccc;
+$text-tertiary: #8e8e8e;
+$text-light: #d4d4d4;
+$text-white: #ffffff;
+$text-disabled: #888;
+
+$accent-primary: #2b76c9;
+$accent-hover: #1e5daa;
+$disabled-bg: #3a3a3a;
+
+// 크기 변수
+$header-height: 4rem;
+$main-header-height: 50px;
+$editor-header-height: 40px;
+$mobile-header-height: 45px;
+
+// 간격 변수
+$spacing-xs: 0.2rem;
+$spacing-sm: 0.5rem;
+$spacing-md: 0.75rem;
+$spacing-lg: 1rem;
+$spacing-xl: 1.5rem;
+
+// 폰트 크기
+$font-xs: 0.75rem;
+$font-sm: 0.8rem;
+$font-base: 0.9rem;
+$font-md: 1.2rem;
+$font-lg: 1.4rem;
+$font-xl: 1.5rem;
+$font-xxl: 1.6rem;
+$font-title: 2.6rem;
+
+// 테두리 반경
+$radius-xs: 2px;
+$radius-sm: 4px;
+$radius-md: 8px;
+$radius-lg: 15px;
+$radius-xl: 30px;
+
+// 믹스인
+@mixin hide-scrollbar {
+  &::-webkit-scrollbar { display: none; }
+  -ms-overflow-style: none;
+  scrollbar-width: none;
+}
+
+@mixin blue-button {
+  padding: $spacing-sm $spacing-lg;
+  color: $text-white;
+  border: none;
+  border-radius: $radius-lg;
+  font-weight: 500;
+  font-size: $font-base;
+  cursor: pointer;
+  transition: background-color 0.2s;
+  white-space: nowrap;
+}
